# Backend/schemas/openapi.yaml
openapi: 3.0.0
info:
  title: Book Database Api
  description: A simple API which provides access to a database of books and their authors.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://turboexhibit-diegosalsa-3030.codio-box.uk/api/v1'
    description: Development server on Codio
components:
  securitySchemes:
    jwtAuth:     
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    NotFoundError:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: JWT is missing or invalid. Please log in using the /login endpoint. 
      content:
        application/json:
          schema:
            type: string
      required:
        - message
    SchemaValidationError:
      description: Invalid resource schema provided in request body.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PermissionError:
      description: Access level not sufficient for accessing requested resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerSideError:
      description: Something went wrong on the server side. If this keeps repeating, contact the admin.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
  parameters:
    page:
      in: query
      name: page
      schema:
        type: integer
        default: 0
      description: The number of the page of items to show. Each page is the size of "limit".
    limit:
      in: query
      name: limit
      schema:
        type: integer
        default: 5
      description: The numbers of items to return.
    order:
      in: query
      name: order
      schema:
        type: string
        default: "id"
      description: The property by which to sort the results.
...
paths:
...
  /book:
    get:
      tags:
        - Book
      summary:
        Get all book records.
      description:
        Get data of all books from the database. Does not include the full contents (request individual books for this).
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/order'
      responses:
        '200':
          description: Array of all book records from the database.
          headers:
            X-Total-Count:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: ./book.json#/definitions/bookGetAll
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      security:
        - jwtAuth: []
      tags:
        - Book
      summary:
        Add book data to the database.
      description:
        Add book data to the database. Returns the record ID.
      responses:
        '201':
          description: Added succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
        '400':
          $ref: '#/components/responses/SchemaValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '500':
          $ref: '#/components/responses/ServerSideError'
      requestBody:
        description: All book data. 
        required: true
        content:
          application/json:
            schema:
              $ref: ./book.json#/definitions/bookAdd
  /book/{id}:
    get:
      tags:
        - Book
      summary:
        Get a single book resource from the database.
      description:
        Get a book resource from the database, together with the full text.
      responses:
        '200':
          description: Full data of the book.
          content:
            application/json:
              schema:
                $ref: ./book.json#/definitions/book
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      security:
        - jwtAuth: []
      tags:
        - Book
      description:
        Modify data of a single book.
      responses:
        '204':
          description: Modified book resource.
        '400':
          $ref: '#/components/responses/SchemaValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerSideError'
      requestBody:
        description: Modifiable book data and book ID.
        required: true
        content:
          application/json:
            schema:
              $ref: ./book.json#/definitions/bookUpd
    delete:
      security:
        - jwtAuth: []
      tags:
        - Book
      description:
        Delete the data of a single book.
      responses:
        '204':
          description: Deleted book.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerSideError'
  /book/unapproved:
    get:
      tags:
        - Book
      summary:
        Get all unapproved book records. Admin only.
      description: >
        Get data of all books from the database which are awaiting approval.
        Does not include the full contents (request individual books for this).
        Admin only.
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/order'
      responses:
        '200':
          description: Array of all unapproved book records from the database.
          headers:
            X-Total-Count:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: ./book.json#/definitions/bookGetAll
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      security:
        - jwtAuth: []
      tags:
        - Book
      summary:
        Change the approval status of a book in the database.
      description: >
        Change the approval status of a book submitted to the database. 
        0 is default, unapproved and invisible to nonadmins.
        Send "approve" to approve the book and make it visible,
        "reject" to delete it from the database.
        Admin only.
      responses:
        '200':
          description: Modified succesfully.
        '400':
          $ref: '#/components/responses/SchemaValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      requestBody:
        description: Book ID and "approve" or "reject".
        required: true
        content:
          application/json:
            schema:
              $ref: ./book.json#/definitions/bookApprove
...
  /review:
    get:
      tags:
        - Review
      summary:
        Get all review records.
      description:
        Get data of all reviews from the database. Does not include the full contents (request individual reviews for this).
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/order'
      responses:
        '200':
          description: Array of all review records from the database.
          headers:
            X-Total-Count:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: ./review.json#/definitions/review
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      security:
        - jwtAuth: []
      tags:
        - Review
      summary:
        Add review data to the database.
      description:
        Add review data to the database. Returns the record ID.
      responses:
        '201':
          description: Added succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
        '400':
          $ref: '#/components/responses/SchemaValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '500':
          $ref: '#/components/responses/ServerSideError'
      requestBody:
        description: All review data. 
        required: true
        content:
          application/json:
            schema:
              $ref: ./review.json#/definitions/reviewAdd
  /review/{id}:
    get:
      tags:
        - Review
      summary:
        Get a single review resource from the database.
      description:
        Get a review resource from the database.
      responses:
        '200':
          description: Full data of the review.
          content:
            application/json:
              schema:
                $ref: ./review.json#/definitions/review
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      security:
        - jwtAuth: []
      tags:
        - Review
      description:
        Modify data of a single review.
      responses:
        '204':
          description: Modified review resource.
        '400':
          $ref: '#/components/responses/SchemaValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerSideError'
      requestBody:
        description: Modifiable review data and review ID.
        required: true
        content:
          application/json:
            schema:
              $ref: ./review.json#/definitions/reviewUpd
    delete:
      security:
        - jwtAuth: []
      tags:
        - Review
      description:
        Delete the data of a single review.
      responses:
        '204':
          description: Deleted review.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerSideError'